requirements.txt
torch>=2.0.0
tensorflow>=2.13.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
PyPDF2>=3.0.0
pdfplumber>=0.9.0
spacy>=3.6.0
transformers>=4.30.0
nltk>=3.8.0
regex>=2023.6.15
click>=8.1.0
pyyaml>=6.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
flask>=2.3.0
werkzeug>=2.3.0
# LLM Integration
openai>=1.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0
requests>=2.31.0  # For OpenRouter support
